Add pagination for get my gifts + add gift count in balance API (#264)
* support pagination for getting gifts

* make gift module independent from giftbiz module

* add gift count in coin exchange balance API

+++ consumer/src/main/kotlin/com/noljanolja/server/consumer/model/User.kt
@@ -25,7 +25,7 @@ data class User(
     val updatedAt: Instant = Instant.now(),
     val referralCode: String = "",
     val referredBy: String = "",
-    var coinBalance: Double = 0.0,
+    var coinBalance: Long = 0,
 )
 
 data class UserPreferences(
